Minor UI changes

+++ client/src/components/Dashboard/Dashboard.js
@@ -6,20 +6,14 @@ import CssBaseline from '@material-ui/core/CssBaseline';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import Avatar from '@material-ui/core/Avatar';
-import AddItem from '../Inventory/AddItem';
-import Stocks from '../Inventory/Stocks';
 import clsx from 'clsx';
 import {
   Container,
   Button,
   ButtonGroup,
-  Grid,
-  Paper,
-  Box,
 } from '@material-ui/core';
 import purple from '@material-ui/core/colors/purple';
 import Map from './Map';
-import auth from '../../auth';
 
 const purp = purple[900];
 const drawerWidth = 241;
@@ -73,10 +67,14 @@ export default function Dashboard(props) {
   const classes = useStyles();
   const fixedHeightPaper = clsx(classes.paper, classes.fixedHeight);
   const logoutUser = async () => {
+    const r = window.confirm("Do you really want to logout?")
+    if(r == true){
     await localStorage.setItem('Token', null);
     await localStorage.setItem('Status', 'LoggedOut');
     await localStorage.setItem('Image', null)
-  };
+    }
+    }
+  
   return (
     <div className={classes.root}>
       <CssBaseline />
